requirements.txt
Django==1.10.4
django-markdown==0.8.4
django-taggit==0.21.3
Markdown==2.6.7
Pillow==3.4.2
